% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bred}
\alias{bred}
\title{bred: Inferring Case-wise Regulatory Networks}
\description{
Network inference methods are computational tools which use
large omics datasets to predict which genes are regulated by which
transcription factors. Since regulatory interactions are context-dependent,
attempting to model regulatory dynamics in the form of a single regulatory
network may have little relevance. This package allows to predict
regulatory networks for individual samples, which can then be clustered
together.
}
